+++ source/utilities/tools.F90
@@ -1265,5 +1265,4 @@ Subroutine get_current_dir(dir_name)
     If (C_ASSOCIATED(c_str_ptr)) Then
       Call c_free(c_str_ptr)
     End If
-
   End Subroutine get_current_dir