+++ docs/user/whatsnew.rst
@@ -1,6 +1,17 @@
 What's New
 ==========
 
+v5.4.7.1 (26 February 2026)
+-------------------------
+
+Changes
+^^^^^^^
+* Added functionaility to write new naming convention for pertinent output datasets (See 5.4.7).
+* Added CF Compliance attributes to output datasets written bed layer thickness, bed layer fractions, and sediment percentiles
+* Now writing CF compliance attributes to each solution's root group instead of "PROPRERTIES" group.
+* Stop writing CF attributes for vector datasets.  Will implement on separated scalars at a later date.
+* Added new function to test a character to see if it is a numeric digit.
+
 v5.4.7 (23 February 2026)
 -------------------------